backend-dev1/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bycrypt]
python-jose[cryptography]
python-dotenv
asyncpg
faker